Se crearon 6 nuevas clases para implementar las subpartes restantes. Ademas, se añadieron ciertos metodos nuevos para facilitar el manejo de las matrices. + A la clase Main.java especificamente al metodo main se añadieron las instancias de cada frame de la practica 10. + Se modificaron cosas en esta clase. No recuerdo xd. + A la clase GraficosCartesianos.java se añadieron nuevos metodos para facilitar el manejo de la multiplicacion de matrices. + Se implementaron y ajustaron a la medida los metodos agregados a la clase GraficosCartesianos.java. + Ademas se renombro el paquete Vectorial a vectorial, esto para que cumpla con las normas de lowercase de java para paquetes. + Se ajustaron algunos metodos, ademas de que se eliminaron unos metodos mal implementados y que fueron reasignados a  CartesianoVectores.

+++ src/graficos/cartesiano/Algoritmico/CartesianoAlgoritmos.java
@@ -86,7 +86,7 @@ public class CartesianoAlgoritmos extends MapaCartesiano {
     }
 
     @Override
-    public int[][] rotar(int[] x, int[] y, int grados){
+    public int[][] rotar(int[] x, int[] y, double grados){
         int[] xR = x.clone();
         int[] yR = y.clone();
